+++ packages/react/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.12.0](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.6...@lubycon/react@1.12.0) (2021-08-14)
+
+
+### Features
+
+* **react:** add useLocalStorage ([#32](https://github.com/Lubycon/lubycon-frontend-libraries/issues/32)) ([72ba173](https://github.com/Lubycon/lubycon-frontend-libraries/commit/72ba1735ab8a5e31b5f958c123b6a21ba3b6dbb3))
+
+
+
+
+
+## [1.11.6](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.5...@lubycon/react@1.11.6) (2021-08-14)
+
+**Note:** Version bump only for package @lubycon/react
+
+
+
+
+
+## [1.11.5](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.4...@lubycon/react@1.11.5) (2021-08-13)
+
+**Note:** Version bump only for package @lubycon/react
+
+
+
+
+
+## [1.11.4](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.3...@lubycon/react@1.11.4) (2021-08-13)
+
+**Note:** Version bump only for package @lubycon/react
+
+
+
+
+
+## [1.11.3](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.2...@lubycon/react@1.11.3) (2021-08-13)
+
+
+### Bug Fixes
+
+* **react:** 새로 작성된 hooks export ([#34](https://github.com/Lubycon/lubycon-frontend-libraries/issues/34)) ([71a4463](https://github.com/Lubycon/lubycon-frontend-libraries/commit/71a4463c296575d4eaa2f9abb6cf41a78fa2f0ea))
+
+
+
+
+
+## [1.11.2](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.1...@lubycon/react@1.11.2) (2021-08-11)
+
+**Note:** Version bump only for package @lubycon/react
+
+
+
+
+
 ## [1.11.1](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/react@1.11.0...@lubycon/react@1.11.1) (2021-08-08)
 
 **Note:** Version bump only for package @lubycon/react