+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/utils@1.9.0...@lubycon/utils@1.9.1) (2021-08-14)
+
+**Note:** Version bump only for package @lubycon/utils
+
+
+
+
+
+# [1.9.0](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/utils@1.8.15...@lubycon/utils@1.9.0) (2021-08-13)
+
+
+### Features
+
+* **utils:** Semantic Version 관련 유틸들 추가 ([#36](https://github.com/Lubycon/lubycon-frontend-libraries/issues/36)) ([2bb5388](https://github.com/Lubycon/lubycon-frontend-libraries/commit/2bb5388efcb2fba08eff15448de718bed2c94646))
+
+
+
+
+
+## [1.8.15](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/utils@1.8.14...@lubycon/utils@1.8.15) (2021-08-13)
+
+**Note:** Version bump only for package @lubycon/utils
+
+
+
+
+
+## [1.8.14](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/utils@1.8.13...@lubycon/utils@1.8.14) (2021-08-11)
+
+**Note:** Version bump only for package @lubycon/utils
+
+
+
+
+
 ## [1.8.13](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/utils@1.8.12...@lubycon/utils@1.8.13) (2021-08-08)
 
 **Note:** Version bump only for package @lubycon/utils